# docs/api-specs/swagger.json
{
  "openapi": "3.0.4",
  "info": {
    "title": "TroLiKOC.API",
    "version": "1.0"
  },
  "paths": {
...
    "/api/payment/sepay-webhook": {
      "post": {
        "tags": [
          "SePayWebhook"
        ],
        "requestBody": {
          "content": {
            "application/json-patch+json": {
              "schema": {
                "$ref": "#/components/schemas/SePayWebhookDto"
              }
            },
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SePayWebhookDto"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/SePayWebhookDto"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/SePayWebhookDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
...
    "/api/Subscription/tiers/{tierId}": {
      "get": {
        "tags": [
          "Subscription"
        ],
        "parameters": [
          {
            "name": "tierId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/SubscriptionTierDto"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SubscriptionTierDto"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/SubscriptionTierDto"
                }
              }
            }
          }
        }
      }
    },
    "/api/Subscription/subscribe/{tierId}": {
      "post": {
        "tags": [
          "Subscription"
        ],
        "parameters": [
          {
            "name": "tierId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
...
    "/api/Wallet/transactions": {
      "get": {
        "tags": [
          "Wallet"
        ],
        "parameters": [
          {
            "name": "page",
            "in": "query",
            "schema": {
              "type": "integer",
              "format": "int32",
              "default": 1
            }
          },
          {
            "name": "size",
            "in": "query",
            "schema": {
              "type": "integer",
              "format": "int32",
              "default": 20
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/TransactionDto"
                  }
                }
              },
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/TransactionDto"
                  }
                }
              },
              "text/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/TransactionDto"
                  }
                }
              }
            }
          }
        }
      }
    },
    "/api/Wallet/topup": {
      "post": {
        "tags": [
          "Wallet"
        ],
        "requestBody": {
          "content": {
            "application/json-patch+json": {
              "schema": {
                "$ref": "#/components/schemas/TopUpRequest"
              }
            },
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TopUpRequest"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/TopUpRequest"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/TopUpRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/WalletDto"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/WalletDto"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/WalletDto"
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
...
      "SePayWebhookDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int64"
          },
          "gateway": {
            "type": "string",
            "nullable": true
          },
          "transactionDate": {
            "type": "string",
            "format": "date-time"
          },
          "accountNumber": {
            "type": "string",
            "nullable": true
          },
          "subAccount": {
            "type": "string",
            "nullable": true
          },
          "code": {
            "type": "string",
            "nullable": true
          },
          "content": {
            "type": "string",
            "nullable": true
          },
          "transferType": {
            "type": "string",
            "nullable": true
          },
          "transferAmount": {
            "type": "number",
            "format": "double"
          },
          "accumulated": {
            "type": "number",
            "format": "double"
          },
          "referenceCode": {
            "type": "string",
            "nullable": true
          },
          "description": {
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
...
      "SubscriptionTierDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "name": {
            "type": "string",
            "nullable": true
          },
          "monthlyPrice": {
            "type": "number",
            "format": "double"
          },
          "maxJobsPerMonth": {
            "type": "integer",
            "format": "int32"
          },
          "maxResolution": {
            "type": "string",
            "nullable": true
          },
          "hasWatermark": {
            "type": "boolean"
          },
          "queuePriority": {
            "type": "string",
            "nullable": true
          },
          "supportsLoRA": {
            "type": "boolean"
          },
          "supportsVoiceCloning": {
            "type": "boolean"
          },
          "isActive": {
            "type": "boolean"
          }
        },
        "additionalProperties": false
      },
      "TopUpRequest": {
        "type": "object",
        "properties": {
          "userId": {
            "type": "string",
            "format": "uuid"
          },
          "amount": {
            "type": "number",
            "format": "double"
          },
          "reference": {
            "type": "string",
            "nullable": true
          },
          "description": {
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
      "TransactionDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "type": {
            "type": "string",
            "nullable": true
          },
          "amount": {
            "type": "number",
            "format": "double"
          },
          "balanceBefore": {
            "type": "number",
            "format": "double"
          },
          "balanceAfter": {
            "type": "number",
            "format": "double"
          },
          "reference": {
            "type": "string",
            "nullable": true
          },
          "description": {
            "type": "string",
            "nullable": true
          },
          "status": {
            "type": "string",
            "nullable": true
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          }
        },
        "additionalProperties": false
      },
      "WalletDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "userId": {
            "type": "string",
            "format": "uuid"
          },
          "balance": {
            "type": "number",
            "format": "double"
          },
          "currency": {
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": false
      }
    }
  },
...
}